:japanese_castle: Add sanity check for stops bulk replace function.

+++ src/api/services/stops-service.ts
@@ -5,6 +5,10 @@ import BaseService from "./base-service"
 
 export class StopsService extends BaseService {
   static async bulkReplace(formId: number, stops: Stop[]): Promise<Stop[]> {
+    if (!stops.every(stop => stop.taid === formId)) {
+      throw new Error('All stops must belong to the same form.');
+    }
+
     return db.transaction(async (transaction) => {
       await transaction("stops").where("taid", formId).delete()
       return transaction("stops").insert(stops).returning("*")
